+++ examples/react-app/lib/evm/custom-chains.ts
@@ -1,0 +1,30 @@
+import type { AppKitNetwork } from "@reown/appkit/networks";
+
+// TODO(refactor): replace with the the network import from `@reown/appkit/networks` when it is added there.
+export const robinhood = {
+  id: 4663,
+  name: "Robinhood",
+  nativeCurrency: {
+    decimals: 18,
+    name: "Ether",
+    symbol: "ETH",
+  },
+  rpcUrls: {
+    default: {
+      http: ["https://rpc.mainnet.chain.robinhood.com"],
+    },
+  },
+  blockExplorers: {
+    default: {
+      name: "Robinhood Chain Blockscout",
+      url: "https://robinhoodchain.blockscout.com",
+      apiUrl: "https://robinhoodchain.blockscout.com/api",
+    },
+  },
+  contracts: {
+    multicall3: {
+      address: "0xca11bde05977b3631167028862be2a173976ca11",
+      blockCreated: 0,
+    },
+  },
+} as const satisfies AppKitNetwork;